Fixed jenkins failure in innodb.innodb_buffer_pool_flushed_stats
Summary: The failure was caused by race conditions in the test case, the counters are updated separately in innodb so race can happen, so the values retrieved might not match, the test case was aware of this so 5% error rate was allowed, but actually it is not enough, when the error rate is more than 5% this test fails. This fix makes sure that most of the dirty pages get flushed before the counters are retrieved so the error rate won't be more than 5%. Also a single page flushing was not counted by the counter of BUF_FLUSH_SINGLE_PAGE. This fix adds it.

Test Plan: jenkins

Reviewers: tianx

Reviewed By: tianx