make gtid rollback safely
Summary:
1) When there are non GTID transactions present
in master's binlog, after slave received some GTID transactions
the non GTID transactions will get skipped because of
slave_gtid_info table.
This is higly unlikely scenario because GTIDs
are generated for every transaction on master but if we
want to rollback GTID deployment by turning off GTIDs
on master, this diff makes sure slaves replicating from that
master are fine by resetting last_gtid seen.

2) When opening a binary log previously the previous_gtid_log_event
was written to binary log only when gtid_mode is ON. But in cases where
GTIDs are rolled back, previous_gtid_log_event when not empty must be
written to store the state of GTIDs so that the gtid_executed won't get
deleted. This ensures that all the slave's gtid_executed is consistent with
master.

3) changing read_only while running transactions by super
users may cause binlog to get corrupted. This is due to the fact
that read_only doesn't block super users and while executing transactions
when THD checks should_write_gtids(), it gets different results
during binlog cache logging and during logging into binary log. cache
the value of should_write_gtids() while initializing binlog cache
to fix this.

Test Plan:
1) mtr test to make sure GTID-less transactions
are not skipped on GTID enabled slave

2) resart a server with gtid_mode=OFF and check previous_gtid_log_events
written to binary log.

3) mysqlslap with 100 connections and toggle read_only from 1 to 0 and
0 to 1. check the binary log is fine in each case using mysqlbinlog.

Reviewers: jtolmer

Reviewed By: jtolmer