Fix SRO + multi-table update statements in replication when logged in statement form.
Summary:
This was contributed as:
https://github.com/facebook/mysql-5.6/pull/11

This is a bug fix for super-read-only not being ignored for the replication thread on multi-table
update statements in sql_parse.cc that would cause replication the stop with the error
"Error 'The MySQL server is running with the --read-only (super) option so it cannot execute this statement'
on query. Default database: 'test'. Query: 'update t2 natural join t1 set b=3'".
The same issue was found in transaction.cc.

To help avoid this issue in the future the logic for testing if read-only should
be enforced is refactored into a new inline function, and existing code that
performed the logic in each location was changed to instead call this function.

Test Plan: mtr test

Reviewers: steaphan, jtolmer

Reviewed By: jtolmer

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1760998

Tasks: 5860804

Signature: t1:1760998:1419977898:a87c4b19a39bc596b16d302da988d879e5176918

Conflicts:
	sql/sql_parse.cc