Fix ha_rocksdb::open to use the correct name
Summary:
The ha_rocksdb::open call constructs the name of the table from the
TABLE argument when attempting to lookup the index in the ddl manager.
However, the names stored in the ddl_manager are generated through the
rocksdb_normalize_tablename() during index creation.

For partitions, the name created from the TABLE arg is different from
the one generated with the normalize call. The "#P#" portion of the name
is needed to lookup the proper index in the ddl_manager.

Differential Revision: https://reviews.facebook.net/D33375