Fix tests of larger InnoDB pages
Summary:

This is from part of original diff that port 32K page support. Since the
original diff was moved to upstream and re-ordered to the front all of
the previously earlier features, it left the test fix part.

Internal:
@tianx: It is not worth the time to break down the test fixes and
sometimes impossible to merge into the earlier features.  So I keep it
as is.

Related original diffs:
https://phabricator.fb.com/D1337667
https://phabricator.fb.com/D1546293
https://phabricator.fb.com/D1608096
https://phabricator.fb.com/D1673586

Reviewers: over, rongrong, steaphan, tianx

Reviewed By: rongrong, steaphan, tianx

Subscribers: mcallaghan, webscalesql-eng@, bakert

Test Plan:
Jenkins
mysqlrpmtest.sh