Increase metadata locks timeout for innodb.innodb-table-online
Summary:
This is another case of lock timeout failure that occasionally happens
on innodb.innodb-table-online. It looks to me the 10 seconds duration is
sometimes not sufficient to acquire lock for the online table rebuild
DDL (alter table) with compression, such that it ended up with timeout
error instead of expected test error. In our failed test, the lock
contention between the DDL and concurrent inserts/updates could lasts
way over 10 seconds.  So I increased the lock_wait_timeout to 300
seconds to be safe.

Test Plan: jenkins

Reviewers: jtolmer

Reviewed By: jtolmer