Correctly handle the max size of DOCUMENT column (16MB - 1).
Summary: the max size of a document column is 16MB - 1. When the size of the FBSON of a JSON document exceeds this limit, currently the document will be truncated (if the sql_mode is not strict), just like what blob does. This behavior doesn't make sense for JSON document. An error should be returned if the column cannot be NULL which will fail the insertion, or set to NULL with warning otherwise (the warning can be treated as an error and fail the insertion eventually if the sql_mode set to strict, say, sql_mode=strict_trans_tables, but this is separate logic)

Test Plan: Jenkins

Reviewers: tianx, santoshb

Reviewed By: santoshb