Increase max user name length to 80
Summary:
Previous bump from 16 to 32 was insufficient, so going all
the way up to 80 this time. 80 is effectively the limit without
making additional changes because of MyISAM's 1000 byte limit on
key length and the primary key of mysql.columns_priv. To compute the
size of the key for that table, the lengths of the char fields are
summed and then multipled by 3, resulting in a limit on the max size
of user name lengths being 81 characters.

Test Plan:
'arc unit --everything'. RPM tests will get
run as part of release process.

Reviewers: santoshb

Reviewed By: santoshb