Improve mtr's support for long temp dirs
Summary:
mtr checks for long temp dir paths and instead creates something
in /tmp. The problem is that a bunch of the code in mtr is inconsistent
in using "$opt_vardir/tmp" rather than the /tmp path. To keep our diff from
upstream smaller, I just added a bit of code to create a symlink between the
2 so that all files still end up in "$opt_vardir/tmp".

Primary motivation for this change is to allow us to correctly capture the
log files from failed xtrabackup tests.

Test Plan:
Created a long directory structure on devserver,
ran ./mtr --verbose, verified that xtrabackup log files end up in
"$opt_vardir/tmp" and that the temp directory in /tmp was correctly
cleaned up.

Reviewers: santoshb

Reviewed By: santoshb