Xtrabackup mtr tests should exit when they hit the real error
Summary:
Xtrabackup mtr tests do a lot of shell, and ignore all the
exit codes. Fix that by adding set -e to all of them.

Test Plan:
./mtr --suite=xtrabackup and verified that tests passed.
Then added a die in innobackupex and verified that we exited at the
exec of the shell script rather than later on a SQL statement.

Reviewers: santoshb

Reviewed By: santoshb