Fix validation of compressed page checksum on flush
Summary:
For http://bugs.mysql.com/bug.php?id=73689, Oracle made a fix to try
and handle pages which have a checksum of zero. Unfortunately, that
fix makes an invalid assumption that the page's LSN being zero means
that the page is empty. That's not a valid assumption to make because
the page cleaner thread was verifying the page checksum prior to
updating the LSN. To fix Oracle's code to correctly handle non-empty
pages that happen to have a computed checksum of zero, change the order
of operations so that the LSN is updated before verifying the checksum.

Test Plan:
./mtr --suite=innodb_stress
Jenkins

Reviewers: santoshb

Reviewed By: santoshb