Store latest direction in Apply_changes_iter
Summary:
When using the Apply_changes_iter class to traverse a key range, we need
to store the latest directional call so that the trx or rdb is pointing
to the appropriate keys. Each time we switch directions, from Next() to
Prev(), or from SeekToLast() to Next(), we need to set trx or rdb to
point at the appropriate keys first.

Differential Revision: https://reviews.facebook.net/D37539

fbshipit-source-id: 88f9ad65ded