Fix insert ... on duplicate key update
Summary:
With unique secondary index support, inserting a record that conflicts
with the secondary key now generates an error. Subsequently, MySQL
attempts to look up the record using the secondary index and these
lookups fail because we always try to match exact.

To fix this, the key lookup should only be performed using the user
defined fields of the secondary index, instead of all fields.

Differential Revision: https://reviews.facebook.net/D37791