Using the default storage engine in the bootstrap
Summary: In order to use the same branch and run same tests for both innodb and rocksdb engines, we will be using the default engine setting to direct which engine to use. So, if you want to run tests against rocksdb, use the following arguments for mtr: --mysqld=--default-storage-engine=rocksdb. If this argument is not provided, innodb bootstrap will be ran and the test will use innodb as default engine.

Test Plan: Ran mtr --start-and-exit for rocksdb, innodb, myisam default engines. Verified that in all cases but rocksdb, innodb was the engine for mysql.slave_XXX tables. For rocksdb storage engine, rocksdb was the engine for these tables.

Reviewers: yoshinorim, hermanlee4, spetrunia, jtolmer

Reviewed By: jtolmer

Differential Revision: https://reviews.facebook.net/D38655