Add feature to FbsonJsonParser and FbsonToJson to parse JSON string with escape sequences
Summary:
To parse the sting with escape chars, I modify the function of parseString.
When the program got the character '\', it means that it's special character
that need to be escaped. We need to treat it specially. And the function
called parseEscape will be called to finish this job.

There are two kinds of escape characters:
 - one is just \b, \f, \n, \r \t and
 - the other is \u which is a Unicode char. The code to parse the unicode is
   also added.

Besides, the way to parse the key is also modified, because Key is also a string.

On the side, we also need to modify the FbsonToJson to make it work with escape
sequences. A function called cstring_to_json is added:
 - when the String in json is printed, it will call this function
 - when Key is printed, this function will also be called.

Test Plan: jenkins, and also added new test cases in json.type_document_innodb

Reviewers: tianx

Reviewed By: tianx