Expose more information to audit plugin
Summary:
We need some extra info for the shadowing and security logging. This is a
simple first step of info that MariaDB actually also exposes.
Now we would have the `query_id` and the database name for general events.
Making as few changes as possible to accomplish it, so I'm just taking the
information from the TDH and exposing it through `mysql_event_general`
struct and as a argument to disconnect.

Test Plan: I didn't find any proper testing for plugin changes, so I'm trusting Jenkins

Reviewers: tianx, chip, santoshb, jtolmer

Reviewed By: chip