Add support for reporting keys/deletes skipped in the extra slow query log
Summary:
The rocksdb perf context stats captures the number of keys and
tombstones skipped during a query. Modify the slow query log expanded
output to capture and report this information back.

Test Plan:
Enable slow query extra and run a few queries that are logged to the
slow query log. Verify the output.

Reviewers: spetrunia, maykov, yoshinorim

Reviewed By: yoshinorim

Differential Revision: https://reviews.facebook.net/D39015