Correctly handle indexes on column prefixes
Summary:
The code used table->field[n] when packing/unpacking index values.
This happens to work most of the time, however the right way to
do it is to use table->key_info[i].key_part[j], which has a field
object that describes the field use in the index.

Test Plan: Added testcase, ran mtr

Reviewers: maykov, jtolmer, yoshinorim, hermanlee4

Reviewed By: hermanlee4

Differential Revision: https://reviews.facebook.net/D39471