Add new function json_contains to check for existence of key-value pair in any level of JSON document
Summary:
This diff introduces the new built-in function: JSON_CONTAINS.

The syntax for the function is:
JSON_CONTAINS (<arg1>, <arg2>, [<arg3>])

- <arg1> could be either a document column or a string literal that represents a JSON document.
- <arg2> is a string that represents a JSON key
- <arg3> is an optional argument that represents a JSON value

This function returns true if the key-value pair represented by <arg2> and <arg3> can be found anywhere in the document <arg1>. This search is not limited to the first level of the JSON object; it will recursively search every element of an array and object for the key-value pair. If <arg3> is omitted, then it will only search for the existence of the key on any arbitrary level.

Depending on the type of <arg3>, it will map to a corresponding JSON type. For example, if <arg3> is NULL, it will search for a JSON null value. If <arg3> is an integer, it will search for a JSON integer value, etc.

Example:
SELECT JSON_CONTAINS('{"k": {"k1":123}}', 'k1')
1

SELECT JSON_CONTAINS('{"k": {"k1":123}}', 'k1', 123)
1

SELECT JSON_CONTAINS('{"k": [123, {"k1": {"k2":"v2"}} ]}', 'k2', 'v2')
1

Errors:
- If <arg1> is not a valid JSON string or valid document column, an ER_INVALID_JSON error will be thrown
- If <arg2> is not a string, an ER_WRONG_ARGUMENTS error will be thrown
- If there are more than 3 arguments or less than 2 arguments, an ER_WRONG_PARAMCOUNT_TO_NATIVE_FCT will be thrown.

Test Plan:
Added test cases for JSON_CONTAINS in:
- json/include/fbson_output_format.inc
- json/include/json_func_common.inc
- json/include/json_func_common_2.inc
- json/include/json_func_common_2_document.inc
- json/t/json_func_literal.test

Reviewers: pengt

Reviewed By: pengt