Add new document item type and DOCUMENT() function
Summary:
This diff introduces a new function DOCUMENT() which takes in a JSON literal string as a parameter and the return type is the document type. Internally, it is represented as a instance of class Item_func_document. This function will validate the JSON literal; if it's not a valid JSON, it will throw a syntax error. If it is a valid JSON, the FBSON will be stored and the function will be a valid expression of document type in SQL.
The difference between this new type and a literal string is that you can add new operators that only make sense for the new document type. It also guarantees that a document item is parsed correctly as valid JSON. Note that all document items must be created using the DOCUMENT() function.

For example, SELECT DOCUMENT('{"k": "v"}') will create a document item successfully where SELECT DOCUMENT('{"k"}') will fail due to invalid JSON.

Test Plan:
New test added:
json.type_document_value.test

Reviewers: pengt

Reviewed By: pengt