Fix order by bug when selecting doc path ordered by doc column
Summary:
This diff fixes an ORDER BY bug in DocStore. The issue was when one tries to
run a query that is ordered by a document column and there is a document path
in the select statement. An example would be:

```
SELECT doc.address.state FROM t ORDER BY doc;
```

This query will actually end up ordering the rows by "doc.address.state" instead
of the overall document column. The reason is because MySQL will attempt to
match items from the SELECT list to items from the ORDER BY list. Since the
field names for doc columns and doc paths are the same, it will assign the item
to that of the doc path.

A fix on doc path comparison was previously introduced on line 7456 of
sql_base.cc. However, an extra condition was added to differentiate between
aliases and document paths. This condition was
"((Item_ident *)find)->parsing_info.num_unresolved_idents > 0". Unfortunately,
document columns will also return FALSE for this condition which causes the
above bug. This diff instead covers all the cases of aliases, document columns,
and document paths in the select/order by list.

- If the item in select list is an alias and is ordered by that alias, then match the
items.
- If the item in select list is a doc path and is ordered by a doc path, then
match the items ONLY if a full doc path comparison is successful
- If the item in select list is a doc path without alias and is ordered by a doc
column, don't match the items as they are completely different types.

Many more tests for ORDER BY were added to cover all cases when ordering by
document columns and document paths.

Test Plan: Added test cases in json.type_document_path

Reviewers: pengt

Reviewed By: pengt