added IF EXISTS to the ALTER TABLE command
Summary:
Added the ability to alter a table only if it exists.  This was
requested for replication purposes.

Test Plan:
The tbl_alter_if_exists test was added.  It attempts to alter
a table that doesn't exist and verifies that it gets a Note but no
error. The it creates the table and again attempts to alter it which now
succeeds with no warning.

Reviewers: hfisk, kknapp, herman, jtolmer

Reviewed By: jtolmer