Adding support updating in FBSON
Summary:
Since FBSON is stored sequentially, so when doing update, the data
sometimes need to be moved. The basic idea is that, when the size of
the document decrease, we do shrink by moving the rest of the data
after the address modified. And then the size of the document
increase, we expand by moving the rest of the data. But some simple
optimizations can be done here:
For updating an integer, we firstly will check whether the integer can
fit into the to be updated FbsonValue. Sometimes, it happens, for example,
we want to update an int64_t to int8_t. For this situation, we will
update it in place.
For updating string, and the length of the string is less than that of
the current one, we update it in place and use tailing zeros to
denotes the end of the string.
All the updating routines are implemented in FbsonUpdater.h and it
maintains a stack to track the path of how to get to the current
node. This is useful, because when the size of the document
changes, we need to update the pack size of all the effected nodes.

Test Plan: Jenkins

Reviewers: pengt, tianx

Reviewed By: tianx