Add supporting of partial update in DocStore
Summary:
This diff is about supporting partial update in DocStore. There are
two types of syntax implemented.

== Implementing of the partial update for traditional syntax ==
The syntax like

  UPDATE table1
  SET col_name.doc_path = new_value|expression [INSERT] [IF [NOT] EXIST]
  WHERE col_name2 = SOMETHING

should be implemented. Here are some examples for it:

  UPDATE table1
  SET col_name1.doc_path = col_name2.doc_path + col_name3.doc_path;

This one is for updating a column in a document path with an expression.

  UPDATE table1
  SET col_name1.doc_path.0 = 123 INSERT

If the col_name1.doc_path is an array, this query will insert integer 123 to the
array. Or that row will not be updated.

  UPDATE table1
  SET col_name1.doc_path = DOCUMENT('{"key":"value"}') IF EXISTS

This one will update the column col_name1.doc_path by assigning an document
to it only when that row has the field specified by the doc_path.

  UPDATE table1
  SET col_name1.doc_path = DOCUMENT('{"key":"value"}') IF NOE EXISTS

This is different with the previous one that the updating will be done only when the
document path specified by doc_path doesn't exist, which means that a new field
will be inserted.

== Implementing of the partial update for function call syntax ==
The syntax like

  SET doc.col.doc_path.fun(arg1, arg2, arg3,...)

should be implemented and which is also mentioned in the design.

Currently, this functions are provided:
- **doc_path.set(new_value | expression)**: assign the given doc_path to the
  value. It can be applied to all JSON values.
- **doc_path.setExists(new_value | expression)**: assign the given doc_path
  to the value when the document path exists. It can be applied to all JSON values.
- **doc_path.setNotExists(new_value | expression)**: assign the given doc_path
  to the value when the document path doesn't exist. It can be applied to all
  JSON values.
- **doc_path.unset()**: delete the element in given doc_path. It can be applied
  to all JSON values.
- **doc_path.inc(value | expression)**:  increase the given element by the value
  in the parameters.
- **doc_path.insertAt(index, value | expression)**: insert an element before the
  index with given value in the parameters.
- **doc_path.insertAtExists(index, value | expression)**: insert an element before
  the index with given value in the parameters, when the index exists.
- **doc_path.insertAtNotExists(index, value | expression)**: insert an element
  before the index with given value in the parameters, when the index doesn't exist.

Here are some examples:

  UPDATE table1
  SET col_name1.doc_path.set(col_name2.doc_path + col_name3.doc_path);

This one is for updating a column in a document path with an expression.

  UPDATE table1
  SET col_name1.doc_path.insertAt(0, 123) INSERT;

If the col_name1.doc_path is an array, this query will insert integer 123 to the
array. Or that row will not be updated.

  UPDATE table1
  SET col_name1.doc_path.setExists(DOCUMENT('{"key":"value"}'));

This one will update the column col_name1.doc_path by assigning an document
to it only when that row has the field specified by the doc_path.

  UPDATE table1
  SET col_name1.doc_path.setNotExists(DOCUMENT('{"key":"value"}'));

Test Plan: Jenkins

Reviewers: pengt, tianx

Reviewed By: tianx