Add support for UNION, UNION ALL, and UNION DISTINCT with document columns and document values
Summary:
This diff adds support for the 3 UNION operators to work with document columns
and document values. This diff does not add support for document paths.

When the UNION operator is invoked, MySQL looks up the type of the left and
right operands and tries to figure out what type the unioned result should be.
However, for the types MYSQL_TYPE_DOCUMENT and MYSQL_TYPE_DOCUMENT_VALUE, there
are no merge rules so it currently crashes if one tries to UNION a doc column.

The fix checks that if one is trying to merge a document, use the merge rule for
a blob type. This diff also adds more comprehensive test cases with UNION
for document columns and DOCUMENT() values.

Test Plan:
Added new test cases in:
- json.type_document_path
- json.type_document_value

Reviewers: pengt

Reviewed By: pengt