Fix backtick support for doc path array index
Summary:
There exists an issue with document paths when an array index with no backquotes
is prefixed by a document key with backquotes. This throws a MySQL syntax error
even though it should be legal. For example, all of the following are legal:
```
select doc.k.0 from t1;
select `doc`.`k`.`0` from t1;
select `doc`.k.0 from t1;
select `doc`.`k`.k from t1;
```

This query will however produce an error:
```
select `doc`.`k`.0 from t1;
```

After a string with backticks, the parsing for the '.' symbol peeks at the next
character to tokenize and if it's a digit, it will be treated as a decimal,
which means .1 will be treated as 0.1

This diff peeks 2 characters back to defer this decimal assumption if that value
was a backtick.

Test Plan: Added test cases in json.type_document_path

Reviewers: pengt

Reviewed By: pengt