Fix XtraBackup confusion between ulong and double
Summary:
In InnoDB srv_max_buf_pool_modified_pct is a double. But the
XtraBackup param initialization code was setting it as a ulong,
resulting in the variable being set to 0 instead of the default
of 50.0. Remove it as an option and let XtraBackup just use the
default value so that we stop getting asserts from buf0flu.cc
about it and srv_max_dirty_pages_pct_lwm.

Test Plan:
./mtr --parallel=auto --suite=xtrabackup and examine logs for the
values of the variables I now print out.
Jenkins
MySQL RPM tests

Reviewers: pengt, tianx, santoshb

Reviewed By: santoshb