Change partial update INC to ignore fields that are not numeric
Summary:
Currently, the partial update function INC will:
- Insert NULL as a JSON value if the key does not exist
- Crash if it is tries to increment the entire document column

This diff will fix the crashes and only insert new incremented values if the
path already exists and it is of a numeric type. Otherwise, it will skip the
update on that row. The reasoning is because for documents, it is unsure whether
all rows for a given document path are numeric, so one might want to increment
all rows that CAN be incremented. If that's the case, the update will need to
ignore all non-numeric rows.

Test Plan: Added test cases in json.type_document_path_partial_update

Reviewers: tianx, santoshb, pengt

Reviewed By: pengt