Expand Field_document to hold document path for DocStore project
Summary:
In this diff, Field_document is used as general class for document path. The document path information used to be
stored in Item_field and doing this makes a lot of things complex, such as every time when the data is need to be
accessed, the parameter of document path has to be passed into. What's more, it makes the implementation of temp table
operations on document column difficult.
In this project, the document path information is stored into Field_document and this makes the code cleaner and
also good support of temp table and the operations on them.
Now we can view a document column as a column space and each document path is a sub set of this space. For example,
doc is a document column and doc.address and doc.children and doc.phone are document paths which are three different
sub sets in this column space. These columns are called virtual column which is generated by the actual document column.
Now, there are actually fields which means the path information should be stored in Field_document instead of the Item_ident.

Test Plan: Jenkins

Reviewers: santoshb, tianx, pengt

Reviewed By: tianx, pengt