Add supporting of document path in view.
Summary:
The reason why view can't work is because of name resolution.
After introducing the generatl document path, this problem can
be fixed simple.
And here a function to convert the name to document path is added,
and the way to do that here is simple and sometimes trick. A better
way is to change the way we view the 'name'.

Test Plan: Jenkins

Reviewers: santoshb, jshao, pengt, tianx

Reviewed By: tianx