Add support of type conversion for document path in order by.
Summary:
Currently, type is unclear for document path in order by, unless the
type information is given in the index and it will always be converted
to string in order by. So when doing order by for int field in the document
path, it will be converted to string and the result will be wrong if
the user want to do order by for int type.

A new syntax is added for it:

  select doc.age order by doc.age as int;

With this, we can tell the MySQL how to do order by for the document path.

And this syntax, if needed can easily be extended to other original column type
in mysql easily.

Test Plan: Jenkins

Reviewers: santoshb, pengt, tianx

Reviewed By: tianx