Fix deprecated use of val_fbson_blob() in SUBDOC function
Summary:
D2318891 uses a val_fbson_blob() which is a function that was removed. D2318891
ran tests before a rebase so this error was not caught.

This diff replaces val_fbson_blob() with val_document_value(buffer) which is the
new way to extract the FbsonValue from an item.

Test Plan:
Ran json.operator_document_subdoc
Will also wait for Jenkins

Reviewers: pengt, tianx

Reviewed By: tianx