Support document path in UNION, UNION DISTINCT, and UNION ALL (includes subquery)
Summary:
Before this diff, document columns will be treated as blob in union. Now a new field
called Field_document path will be created for the temp table created by UNION.
With this, we can support document paths with the different UNION operators.

This involves adding a new set of merge rules between two different field types in
the UNION. This diff also supports a document path in the outer query where the
inner query is a UNION. However, this is only supported if all the UNION'ed fields
are of type DOCUMENT (see below for examples)

This is valid:
```
SELECT doc.a FROM (
   (SELECT doc FROM t1) UNION (SELECT DOCUMENT('{"a":123}'))
) alias;
```

This is not because the second select in the inner query is a string:
```
SELECT doc.a FROM (
   (SELECT doc FROM t1) UNION (SELECT '{"a":123}')
) alias;
```

Test Plan: Added test cases in json.type_document_path

Reviewers: santoshb, eason, pengt, tianx

Reviewed By: tianx