Fix type checking for SIMILAR/SUBDOC/CONTAINS inside HAVING
Summary:
When SIMILAR/SUBDOC/CONTAINS is used inside a HAVING, an argument error is thrown
because the doc path is a reference item instead of a field item. Because of
this, it fails the check for valid document type. This diff extends the check
to reference items similar to how cached items are handled.

Test Plan:
Jenkins, more test cases in:
- json.operator_document_similar
- json.operator_document_subdoc
- json.operator_document_contains

Reviewers: santoshb, pengt, tianx

Reviewed By: tianx