Fixed the drop table binlog encoding errors
Summary: Fixed the issue that drop table executed on master but failed on slaves(encoding issue). The issue is due to the fact that the encoding is not set correctly when preparing the binlog events of drop tableS. The binlog is therefore written with the wrong encoding, which hinders the slave from replicating. This diff fixed the encoding issue for all the potential scenarios for drop tables.

Test Plan:
mysqltest.sh --valgrind --repeat=32 --parallel=32 droptable_encoding
The test case covers all the code paths of this modification.

Reviewers: santoshb

Reviewed By: santoshb