Improving stats collection
Summary:
Changing the stats computation logic to adjust stats every time a new sst file is flushed or whenever compaction is performed.
Changes in stats are stored into data dictionary inside a separate myrocks thread.
This is not a complete work. Here is what's left and will be done in subsequent diffs:
- analyze table implementation (recompute stats for a particular table)

Test Plan: ran stats-specific mtr tests for rocksdb

Reviewers: yoshinorim, hermanlee4

Subscribers:

Differential Revision: https://reviews.facebook.net/D48495