Add support for client aliases
Summary:
Add support for client-side aliases with the \a command.

These are read from a configurable directory (specified with client-alias-dir
in the [client] section, or on the command line with --client-alias-dir).  Each
file within that directory will be read; lines containing an "=" will be
treated as "alias=command" pairs, and those starting with a "#" will be ignored
as comments.

Aliases work in an interesting way, thanks to MySQL's use of readline.  Aliases
starting with a "!" are run immediately; so you might have this in an alias
file:

# This takes forever to type the old way!
trans = !SHOW ENGINE InnoDB TRANSACTION STATUS\G

Now, in a MySQL session, you can just type:

\a trans

And you'll get the output of that statement.  In fact, you can include anything
you'd like there, even multiple statements or other MySQL client commands;
those who find themselves changing the PAGER to search for some particular
string might choose to do something like:

trumpify = !\P grep -iE 'trump|the Donald'

which gives you the ability to see:

mysql> \a trumpify
PAGER set to 'grep -iE 'trump|the Donald''

Now, the other case to discuss is what happens when you don't have a "!" at the
beginning.  If anything, this is more interesting, because it works kind of
like expandtext.  It's hard to explain in static text, but here's a try:

Let's say you have your prompt:

mysql>

You've already loaded your aliases file, which contains this line:

sct=SHOW CREATE TABLE ``\G

At your prompt, you type '\a sct' and hit enter.  Here's what happens:

mysql> \a sct
mysql> SHOW CREATE TABLE ``\G

Your cursor is now sitting at the end of the expansion of the alias, but you're
still in readline mode, meaning that you can edit the line however you'd like
before you hit enter.  In this case, that means you can put in the name of the
table you'd like to investigate simply by using your arrow keys.  Of course,
you could also set up your alias like:

sct=SHOW CREATE TABLE `

in which case, when you hit enter on '\a sct', you'd just have to supply the
table name and an appropriate ending.

==Security==
Aliases are disabled in batch mode.  A failed alias substitution is immediately
fatal (since you might have been relying on an alias that you've just pasted in
or otherwise gotten us to process before going on to the next operation).

Alias directories should be treated with the same care applied to bashrc and
other, similar files.  Aliases do not ask for confirmation before running, and
can do *anything your client can do*!  Now, that said, they only happen if you
type '\a' first, so you're unlikely to accidentally trigger one.

Currently, first-match is used, which means that if you have:

trans = !SHOW ENGINE InnoDB TRANSACTION STATUS\G

you can run it by just typing '\a t'.  Note that this may lead to unexpected
behavior if someone defined another command starting with "t", and because
file sort order inside the directory cannot be guaranteed, you can only trust
ordinality of aliases being loaded *inside* the same file (and probably should
not trust that ordinality, at all, if you're truly being careful).  Basically,
shortcutting aliases is a feature that was easy to add, so I did, but don't
use it unless you're very sure of what you're doing.

==Future==
Similar to MySQL, if additional alias types are desired ('\a sct tablename'
currently runs the alias named 'sct tablename', but wouldn't it be neat if it
used the arg?), the alias file format will require them to be embedded in
comments for backwards compatibility.

The ability to run aliases without '\a' was considered, but rejected as too
dangerous.  Ask nicely and defend your reasoning and it could yet come back.

Test Plan:
I tested in a few common scenarios.  P20028095 is an example aliases
file.  There are probably several bugs, but they *should* be limited to the
aliases implementation, and not be (easy to) trigger from normal use of the
client.

Reviewers: jtolmer

Differential Revision: https://reviews.facebook.net/D47163